Initial migration from S3 to DynamoDB for locking
This breaks quite a lot of tests due to changes to the interface.
We will also require a test / mocking strategy for this new
DynamoDB implementation. Rather than using granular S3 get, and put
operations which can be easily mocked with an in-memory implementation,
we are now relying on DynamoDB's expressions, and conditions.

Affected tests:

- release_lock_test.go
- integration_test.go