Preserve existing locks, and introduce DynamoDB locking as an additional lock
We want existing locks to be respected as various callers / systems
may continue relying on older versions of `GrabLocks` which uses S3.
Thus, if we replaced it, we would not be accounting for legitimate
locks being placed in S3.

Instead, this commit adds the DynamoDB locking as an additional
lock (layer). Eventually, we can replace the S3 locking entirely
once all callers / systems are migrated to using the latest `step`
with DynamoDB.