Set the lock table name using the lambda name, taken from the context
To avoid breaking uses of `bifrost.Release` by introducing a new
field (as per CR feedback), we are now passing the lock table name
after generating it from the context of the lambda execution.
In the handlers, by default we have postfixed `-locks` to the lock
table name (i.e. `<lambda name>-locks`).

This commit also improves the integration test with DynamoDB by simply
capturing all the requests made, and checking if they are as expected.