Remove S3 root lock, and rely on DynamoDB locking
We recently introduced DynamoDB for root locking (a93d5f02),
but we did not remove the use of S3 for root locking.
The reason for this was so that we can gracefully migrate all deployers,
otherwise there will be some deployers (in an environment) relying on
DynamoDB, and some deployers relying on S3 which may result in
some overlaps not being caught.

Now that the deployers have been migrated over, we can remove
the use of S3 for root locking. We will still be using it for
user locking, and release locking as both of those are less likely
to run into race conditions, and the release lock is more of an
idempotency key which we never remove / clean up.

The `Test_DeployHandler_Execution_Errors_Root_LockError` test
was removed as we should never reach a state where we have
acquired a lock, but with an error in the process of doing so.
We may have future implementations that does this, but in terms
of integration testing, our DynamoDB implementation will either
succeed in acquiring a lock without any error or fail.
The `LockError` condition will not be triggered.

The `assertNoRootLockWithReleseLock`, and `assertNoRootLockNoReleseLock`
were also replaced with `assertHasReleaseLock`, `assertNoReleaseLock`,
and checks on the lock items that were added, and/or removed from DynamoDB.